requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
django-filter
djangorestframework-simplejwt
gunicorn
reportlab
openpyxl
Pillow

